More fixes
Conflicts:
	src/Hanzo/Bundle/ServiceBundle/Services/DeadOrderService.php

+++ src/Hanzo/Bundle/ServiceBundle/Services/DeadOrderService.php
@@ -55,6 +55,10 @@ class DeadOrderService
         $this->debug  = $debug;
 
         $this->debug("Starting DeadOrderService Auto Clean");
+        if ( $this->dryrun )
+        {
+          $this->debug("Dry run mode");
+        }
 
         $toBeDeleted = array();
         $toBeDeleted = $this->getOrdersToBeDeleted();
@@ -80,6 +84,7 @@ class DeadOrderService
             $status = $this->checkOrderForErrors($order);
             if ( $status['is_error'] )
             {
+                $this->debug("Order queued to be deleted (".$order->getId().")");
                 $toBeDeleted[] = $order;
             }
         }
@@ -201,7 +206,16 @@ class DeadOrderService
                     }
                     else
                     {
-                      $this->debug( "Not dispatching order.payment.collected event (dryrun)" );
+                      if ( !$this->dryrun )
+                      {
+                        $order->save();
+                        $this->debug( "Dispatching order.payment.collected event" );
+                        $this->eventDispatcher->dispatch('order.payment.collected', new FilterOrderEvent($order));
+                      }
+                      else
+                      {
+                        $this->debug( "(Dryrun) Dispatching order.payment.collected event" );
+                      }
                     }
                     break;
 
@@ -225,8 +239,15 @@ class DeadOrderService
     {
         foreach ($toBeDeleted as $order) 
         {
-            $this->debug("Deleting order: ".$order->getId());
-            $order->delete();
+            if ( !$this->dryrun )
+            {
+              $this->debug("Deleting order: ".$order->getId());
+              $order->delete();
+            }
+            else
+            {
+              $this->debug("(Dryrun) Deleting order: ".$order->getId());
+            }
         }
     }
